# Introduces write-coalescing to FST algorithm